move mass_density to plasmapy.formulary.misc
